Upload File Example documentation update.
Changed the uploadFiles options object to use **progressCallback** instead of **progress** because that is how it's named in FS.common.js.  This tripped me up until I dug into the source to figure out what the callback is actually named.

Note that the downloadFile option is named **progress** both the README example and the FS.common.js file.  Not sure if it's better to rename the uploadFiles option to match downloadFile option or not.

Thanks, this is a great extension to React Native.